refactor(npm): move resolver logic to resolver.py
Resolves https://github.com/hermetoproject/hermeto/issues/1156

Move npm dependency resolve/download/update workflow from
npm/main.py into npm/resolver.py.

Keep compatibility imports in npm/main.py so existing call sites and
tests remain stable during this commit series.

No behavior changes.
